+++ play_game.rb
@@ -1,0 +1,6 @@
+require_relative 'V1/characters'
+require_relative 'V1/helpers'
+require_relative 'V1/game_engine'
+
+V1::GameEngine.play_game
+puts 'yo'
